detect/tcp: rename tcp keyword files
