+++ src/detect-engine-register.c
@@ -75,12 +75,12 @@
 #include "detect-base64-decode.h"
 #include "detect-base64-data.h"
 #include "detect-ipopts.h"
-#include "detect-flags.h"
+#include "detect-tcp-flags.h"
 #include "detect-fragbits.h"
 #include "detect-fragoffset.h"
 #include "detect-gid.h"
-#include "detect-ack.h"
-#include "detect-seq.h"
+#include "detect-tcp-ack.h"
+#include "detect-tcp-seq.h"
 #include "detect-content.h"
 #include "detect-uricontent.h"
 #include "detect-pcre.h"
@@ -106,7 +106,7 @@
 #include "detect-msg.h"
 #include "detect-rev.h"
 #include "detect-flow.h"
-#include "detect-window.h"
+#include "detect-tcp-window.h"
 #include "detect-ftpbounce.h"
 #include "detect-isdataat.h"
 #include "detect-id.h"
