+++ src/detect-tcp-window.h
@@ -15,26 +15,18 @@
  * 02110-1301, USA.
  */
 
-/**
- * \file
- *
- * \author Brian Rectanus <brectanu@gmail.com>
- */
+#ifndef __DETECT_WINDOW_H__
+#define __DETECT_WINDOW_H__
 
-#ifndef __DETECT_SEQ_H__
-#define __DETECT_SEQ_H__
+#define MIN_WINDOW_VALUE 0
+#define MAX_WINDOW_VALUE 65535
 
-/**
- * \brief seq data
- */
-typedef struct DetectSeqData_ {
-    uint32_t seq;                    /**< seq to match */
-} DetectSeqData;
+typedef struct DetectWindowData_ {
+    uint8_t negated;	/** negated? 1=True : 0=False */
+    uint16_t size;     /** window size to match */
+} DetectWindowData;
 
-/**
- * \brief Registration function for ack: keyword
- */
-void DetectSeqRegister(void);
+/* prototypes */
+void DetectWindowRegister (void);
 
-#endif /* __DETECT_SEQ_H__ */
-
+#endif /* __DETECT_WINDOW_H__ */
